Binary expression of cell existence

as_cuboid() expand a coordinate set to a binary cuboid with zero-padding. The data.frame representation is ordered in the same way as the array representation. They have some extra attributes to ensure round-trip conversion.

Usage

as_cuboid(x, expand = 0L)

# S3 method for class 'cuboid'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

# S3 method for class 'cuboid'
as.array(x, ...)

Arguments

x

a data.frame with (x, y, z) or 3D array.

expand

an integer.

row.names, optional, ...

ignored; to suppress warnings for S3 methods.