Utilities for dispersion summary_row() and mutate_chisq() calculate various dispersion stats. stats_dispersion() is a shortcut to call them both at once.

Utilities for dispersion summary_row() and mutate_chisq() calculate various dispersion stats. stats_dispersion() is a shortcut to call them both at once.

Usage

stats_dispersion(x)

summary_row(x)

mutate_chisq(.data, ...)

rpois_dispersion(n, lambda, nrep = 1L)

tibble_dchisq(x, n)

Arguments

x

A numeric vector.

.data

A data.frame.

...

passed to dplyr::mutate().

n, lambda

passed to stats::rpois().

nrep

number of replications.

Details

rpois_dispersion() summarizes the results of stats::rpois().

tibble_dchisq() generates a tibble.